requirements.txt
streamlit
python-docx
fastapi
uvicorn
anthropic
python-dotenv
pydantic
